0001112: Categorias Selecionadas

+++ models/search/search_model.php
@@ -104,6 +104,9 @@ public function add($data) {
     public function update($data) {
         $collection_id = $data['collection_id'];
         if ($data['property_id'] != '') {
+            if($data['property_id']==get_term_by('slug', 'socialdb_property_fixed_tags', 'socialdb_property_type')->term_id){
+                $data['property_id'] = 'tag';
+            }
             $facets = get_post_meta($collection_id, 'socialdb_collection_facets');
             if($facets  &&  is_array($facets) && !in_array($data['property_id'], $facets)){
                 add_post_meta($collection_id, 'socialdb_collection_facets', $data['property_id']);
@@ -151,7 +154,11 @@ public function add($data) {
     }
 
     public function delete($data) {
-        delete_post_meta($data['collection_id'], 'socialdb_collection_facets', $data['facet_id']);
+        if($data['facet_id']==='tag'){
+            $data['facet_id'] = get_term_by('slug', 'socialdb_property_fixed_tags', 'socialdb_property_type')->term_id;
+        }
+        if($data['facet_id'])
+            delete_post_meta($data['collection_id'], 'socialdb_collection_facets', $data['facet_id']);
 
         $result['title'] = __('Success','tainacan');
         $result['msg'] = __('Facet successfully deleted.','tainacan');
